+++ package.json
@@ -36,6 +36,7 @@
     "class-transformer": "^0.5.1",
     "class-validator": "^0.13.2",
     "crypto-js": "^4.1.1",
+    "faker": "5.5.3",
     "figlet": "^1.5.2",
     "gradient-string": "^2.0.0",
     "mongoose": "^6.2.0",
