+++ src/app.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { MongooseModule } from '@nestjs/mongoose';
 import securityConfig from '@src/config/security.config';
-import { FindMeSecurityModule } from '@src/modules/find-me/security/find-me-security.module';
-import { FindMeUsersModule } from '@src/modules/find-me/users/find-me-users.module';
+import { FindMeSecurityModule } from '@src/modules/find-me-security/find-me-security.module';
 import mongodbConfig from './config/mongodb.config';
+import { FindMeUsersModule } from '@src/modules/find-me-users/find-me-users.module';
+import { FindMeSeederModule } from '@src/modules/find-me-seeder/find-me-seeder.module';
 
 @Module({
   imports: [
@@ -18,6 +19,8 @@ import mongodbConfig from './config/mongodb.config';
     MongooseModule.forRoot(mongodbConfig().mongodb.uri),
     FindMeUsersModule,
     FindMeSecurityModule,
+    FindMeSeederModule,
   ],
 })
-export class AppModule {}
+export class AppModule {
+}
