+++ src/modules/find-me-seeder/find-me-seeder.module.ts
@@ -1,0 +1,5 @@
+import { Module } from '@nestjs/common';
+import { FindMeSeederService } from '@src/modules/find-me-seeder/find-me-seeder.service';
+
+@Module({ providers: [ FindMeSeederService ] })
+export class FindMeSeederModule {}
