+++ src/modules/find-me-seeder/find-me-seeder.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class FindMeSeederService {
+  public async seedFindMeUsers(amount: number): Promise<void> {
+
+  }
+}
