+++ src/modules/find-me-users/exceptions/find-me-user-already-exist.exception.ts
