+++ src/modules/find-me-users/find-me-users.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { FindMeSecurityModule } from '@src/modules/find-me/security/find-me-security.module';
-import { FindMeUsersController } from '@src/modules/find-me/users/find-me-users.controller';
-import { FindMeUsersService } from '@src/modules/find-me/users/find-me-users.service';
-import { FindMeUser, FindMeUserSchema } from '@src/modules/find-me/users/schemas/find-me-user.schema';
+import { FindMeSecurityModule } from '@src/modules/find-me-security/find-me-security.module';
+import { FindMeUser, FindMeUserSchema } from '@src/modules/find-me-users/schemas/find-me-user.schema';
+import { FindMeUsersController } from '@src/modules/find-me-users/find-me-users.controller';
+import { FindMeUsersService } from '@src/modules/find-me-users/find-me-users.service';
 
 @Module({
   imports: [
