+++ src/modules/find-me-users/find-me-users.service.ts
@@ -1,10 +1,10 @@
 import { ConflictException, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { FindMeSecurityService } from '@src/modules/find-me/security/find-me-security.service';
-import { CreateFindMeUserDto } from '@src/modules/find-me/users/dto/create-find-me-user.dto';
-import findMeUserAlreadyExistException from '@src/modules/find-me/users/exceptions/find-me-user-already-exist.exception';
-import { FindMeUser, FindMeUserDocument } from '@src/modules/find-me/users/schemas/find-me-user.schema';
+import { FindMeSecurityService } from '@src/modules/find-me-security/find-me-security.service';
 import { Model } from 'mongoose';
+import { CreateFindMeUserDto } from '@src/modules/find-me-users/dto/create-find-me-user.dto';
+import { FindMeUser, FindMeUserDocument } from '@src/modules/find-me-users/schemas/find-me-user.schema';
+import findMeUserAlreadyExistException from '@src/modules/find-me-users/exceptions/find-me-user-already-exist.exception';
 
 @Injectable()
 export class FindMeUsersService {
